frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["PatStLouis <patrick.st-louis@opsecid.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = {extras = ["ext"], version = "^3.0.3"}
github-flask = "^3.2.0"
flask-session = "^0.8.0"
flask-cors = "^5.0.0"
flask-qrcode = "^3.2.0"
# redis-py-cluster
redis = ">=3.0.0,<4.0.0"
flask-security = {extras = ["mfa"], version = "^5.5.2"}
aries-askar = "^0.3.2"
py-webauthn = "^0.0.4"
pydantic = "^2.9.2"
# https://github.com/wbond/oscrypto/issues/78
oscrypto = { git = "https://github.com/wbond/oscrypto.git", rev = "1547f53" }
httpx = "^0.27.2"
flask-rq = "^0.2"
redis-py-cluster = "^2.1.3"
setuptools = "^75.2.0"
flask-limiter = {extras = ["redis"], version = "^3.8.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
